Move BuildFileRetriever into BuildFileRetrievers
Thus it is becoming the base class for BFR
